backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
PyPDF2
faiss-cpu
sentence-transformers
transformers
torch
